+++ wezterm-font/src/lib.rs
@@ -460,6 +460,7 @@ struct FontConfigInner {
     built_in: RefCell<Arc<FontDatabase>>,
     title_font: RefCell<Option<Rc<LoadedFont>>>,
     pane_select_font: RefCell<Option<Rc<LoadedFont>>>,
+    char_select_font: RefCell<Option<Rc<LoadedFont>>>,
     fallback_channel: RefCell<Option<Sender<FallbackResolveInfo>>>,
 }
 
@@ -479,6 +480,7 @@ impl FontConfigInner {
             metrics: RefCell::new(None),
             title_font: RefCell::new(None),
             pane_select_font: RefCell::new(None),
+            char_select_font: RefCell::new(None),
             font_scale: RefCell::new(1.0),
             dpi: RefCell::new(dpi),
             config: RefCell::new(config.clone()),
@@ -495,6 +497,7 @@ impl FontConfigInner {
         fonts.clear();
         self.title_font.borrow_mut().take();
         self.pane_select_font.borrow_mut().take();
+        self.char_select_font.borrow_mut().take();
         self.metrics.borrow_mut().take();
         *self.font_dirs.borrow_mut() = Arc::new(FontDatabase::with_font_dirs(config)?);
         Ok(())
@@ -630,6 +633,22 @@ impl FontConfigInner {
         let loaded = self.make_title_font_impl(myself, config.window_frame.font_size)?;
 
         title_font.replace(Rc::clone(&loaded));
+
+        Ok(loaded)
+    }
+
+    fn char_select_font(&self, myself: &Rc<Self>) -> anyhow::Result<Rc<LoadedFont>> {
+        let config = self.config.borrow();
+
+        let mut char_select_font = self.char_select_font.borrow_mut();
+
+        if let Some(entry) = char_select_font.as_ref() {
+            return Ok(Rc::clone(entry));
+        }
+
+        let loaded = self.make_title_font_impl(myself, Some(config.char_select_font_size))?;
+
+        char_select_font.replace(Rc::clone(&loaded));
 
         Ok(loaded)
     }
@@ -990,6 +1009,10 @@ impl FontConfiguration {
 
     pub fn pane_select_font(&self) -> anyhow::Result<Rc<LoadedFont>> {
         self.inner.pane_select_font(&self.inner)
+    }
+
+    pub fn char_select_font(&self) -> anyhow::Result<Rc<LoadedFont>> {
+        self.inner.char_select_font(&self.inner)
     }
 
     /// Given a text style, load (with caching) the font that best